Optimization of forms for countries
Adjusting forms creation and management of countries according to customer requirements
