--- src/components/apis/crudCountry/CountryCreationForm.css
@@ -1,102 +1,0 @@
-body {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    height: 100vh;
-    margin: 0;
-    font-family: "Arial", sans-serif;
-    background-color: #f4f4f4;
-  }
-  
-  .creation--country {
-    display: flex;
-    background-color: #ffffff;
-    flex-wrap: wrap;
-    justify-content: space-around;
-    width: 100%;
-    max-width: 800px;
-    margin: 20px;
-    padding: 20px;
-    box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
-    border-radius: 8px;
-  }
-  .country--code {
-    display: flex;
-    flex-direction: column;
-    align-items: center; /* Centra horizontalmente */
-    justify-content: center; /* Centra verticalmente */
-    width: 30%;
-    margin-bottom: 20px;
-    margin-right: 25px;
-  }
-  .obtain-information {
-    display: flex;
-    flex-direction: column;
-    width: 48%; /* Ajuste el ancho según sea necesario */
-    margin-bottom: 20px;
-  }
-  
-  label {
-    font-weight: bold;
-    text-align: left;
-    margin-bottom: 5px;
-  }
-  
-  input[type="text"] {
-    width: 100%;
-    padding: 10px;
-    font-size: 16px;
-    border: 1px solid #ccc;
-    border-radius: 4px;
-    margin-bottom: 10px;
-    box-sizing: border-box;
-  }
-  
-  .btn-consult {
-    padding: 10px 20px;
-    margin: 1.5em;
-    font-size: 16px;
-    background-color: #00745e;
-    color: white;
-    border: none;
-    border-radius: 6px;
-    cursor: pointer;
-    transition: background-color 0.3s;
-  }
-  
-  button {
-    padding: 10px 20px;
-    margin: 1.5em;
-    font-size: 16px;
-    background-color: #4caf50;
-    color: white;
-    border: none;
-    border-radius: 4px;
-    cursor: pointer;
-    transition: background-color 0.3s;
-  }
-  .btn--clear {
-    background-color: #e74c3c;
-  }
-  
-  .btn-create {
-    background-color: #2ecc71;
-  }
-  button:hover {
-    background-color: #403e3e;
-  }
-  
-  .obtain-information-tow {
-    display: flex;
-    flex-wrap: wrap;
-    justify-content: space-between;
-    width: 100%;
-  }
-  
-  @media (max-width: 600px) {
-    .country--code,
-    .obtain-information {
-      width: 100%;
-    }
-  }
-
